+++ app/src/main/java/ru/practicum/android/diploma/data/vacancy/SearchVacanciesNetworkDataSource.kt
@@ -33,28 +33,4 @@ class SearchVacanciesNetworkDataSource(
             }
         )
     }
-
-    suspend fun getVacancyDetails(id: String): ApiResponse<VacancyDetails> {
-        return networkClient.execute {
-            val dto = hhApi.getVacancyDetails(id)
-            formatDetails(dto)
-        }
-    }
-
-    private fun formatDetails(dto: VacancyDetailsDto): VacancyDetails {
-        return VacancyDetails(
-            id = dto.id,
-            title = dto.name,
-            employerName = dto.employer?.name ?: "",
-            logoUrl = dto.employer?.logoUrls?.size90,
-            salaryFrom = dto.salary?.from,
-            salaryTo = dto.salary?.to,
-            salaryCurrency = dto.salary?.currency,
-            experience = dto.experience?.name,
-            employment = dto.employment?.name,
-            schedule = dto.schedule?.name,
-            descriptionHtml = dto.description ?: "",
-            keySkills = dto.keySkills?.mapNotNull { it.name } ?: emptyList()
-        )
-    }
 }
