+++ app/src/main/java/ru/practicum/android/diploma/domain/vacancy/api/VacancyDetailsRepository.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.domain.vacancy.api
+
+import ru.practicum.android.diploma.data.network.ApiResponse
+import ru.practicum.android.diploma.domain.vacancy.models.VacancyDetails
+
+interface VacancyDetailsRepository {
+    suspend fun getVacancyDetails(id: String): ApiResponse<VacancyDetails>
+}
