+++ app/src/main/res/layout/fragment_vacancy.xml
@@ -12,7 +12,7 @@
         android:layout_height="match_parent"
         android:text="Vacancy"
         android:gravity="center"
-        android:textSize="60sp"
+        android:textSize="12sp"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent"/>
